Created a LoadoutManager component which receives input and re-directs it to the appropriate loadout.  Will control most general loadout tasks, such as checking if alive, swapping loadouts, and adding the loadout components to the player on initialization.
Created ModelLoadout component, which is a template for future loadouts, in order to insure they interface with the LoadoutManager.

Created StatusManager, intended to be very similar to the StatManager in prototype, except this time it should control all critical status information, as opposed to storing data across a mess of different components.  If you ever need to know the status of an object, the StatusManager has the final say.

Couple more improvements to player controller, although rotation currently has me frustrated, so I made the other things.